calendar

+++ frontend/src/app/(private)/clients/page.tsx
@@ -1,8 +1,13 @@
 "use client";
+import Input from "@/components/input/input";
 import Register from "@/components/register";
 import useApiPrivate from "@/hooks/apiPrivate";
 import { queryClient } from "@/hooks/queryClient";
-import { Clients, useClients as useClientsHook } from "@/hooks/useClients";
+import {
+  Clients as Client,
+  useClients as useClientsHook,
+} from "@/hooks/useClients";
+import { ChangeEvent, useState } from "react";
 import {
   Modal,
   ModalContent,
@@ -11,14 +16,22 @@ import {
   ModalFooter,
   Button,
 } from "@nextui-org/react";
-import { useState } from "react";
+import Link from "next/link";
 
 const useClients = () => {
-  const [itemDelete, setItemDelete] = useState<Clients | null>(null);
-  const handleItemDelete = (newItem: Clients | null) => setItemDelete(newItem);
+  const [itemDelete, setItemDelete] = useState<Client | null>(null);
+  const [nameConfirmDelete, setNameConfirmDelete] = useState<string>("");
+  const onChangeNameConfirmDelete = (e: ChangeEvent<HTMLInputElement>) =>
+    setNameConfirmDelete(e.target.value);
+  const handleItemDelete = (newItem: Client | null) => setItemDelete(newItem);
+
   const api = useApiPrivate();
 
-  const deleteItem = async (id: number) => {
+  const deleteItem = async (item: Client) => {
+    if (nameConfirmDelete !== item.firstName) {
+      return;
+    }
+    const { id } = item;
     await api.delete(`/clients/delete/${id}`);
     queryClient.invalidateQueries(["clients"]);
     handleItemDelete(null);
@@ -27,20 +40,27 @@ const useClients = () => {
   return {
     itemDelete,
     handleItemDelete,
+    onChangeNameConfirmDelete,
+    nameConfirmDelete,
     deleteItem,
   };
 };
 
 export default function Clients() {
-  const { itemDelete, handleItemDelete, deleteItem } = useClients();
-  const { clients }: { clients: Clients[] | undefined } =
-    useClientsHook().getClients();
+  const {
+    itemDelete,
+    handleItemDelete,
+    deleteItem,
+    onChangeNameConfirmDelete,
+    nameConfirmDelete,
+  } = useClients();
+  const { clients } = useClientsHook().getClients();
 
   return (
     <>
       <div className="w-auto flex min-w-[40rem] flex-col gap-2">
         <div className="text-xl text-cyan-600 mt-5">Todos os meus clientes</div>
-        {clients?.map((item: Clients) => (
+        {clients?.map((item: Client) => (
           <Register.Root>
             <Register.Compartiment>
               <Register.Title>Primeiro nome</Register.Title>
@@ -50,12 +70,30 @@ export default function Clients() {
               <Register.Title>Email</Register.Title>
               <Register.Content>{item?.email || "-"}</Register.Content>
             </Register.Compartiment>
+            <Register.Compartiment>
+              <Register.Title>Telefone</Register.Title>
+              <Register.Content>{item.phone || '-'}</Register.Content>
+            </Register.Compartiment>
             <Register.Compartiment className="flex-none flex-row">
               <Register.ButtonTrash onClick={() => handleItemDelete(item)} />
               <Register.ButtonEdit />
             </Register.Compartiment>
           </Register.Root>
         ))}
+        {!clients ||
+          (clients?.length < 1 && (
+            <div className="flex flex-col text-lg opacity-70">
+              <div>Ainda não tem nenhum Cliente Cadastrado</div>
+              <div>
+                <Link
+                  href={"clients/create"}
+                  className="text-cyan-500 opacity-80 hover:opacity-100"
+                >
+                  Cadastre seu primeiro cliente!
+                </Link>
+              </div>
+            </div>
+          ))}
       </div>
       <Modal
         onOpenChange={() => handleItemDelete(null)}
@@ -68,8 +106,20 @@ export default function Clients() {
             Excluir registro
           </ModalHeader>
           <ModalBody>
-            <p>Deseja realmente excluir o registro {itemDelete?.firstName}?</p>
-            <p>Depois de excluir, não haverá como recupera-lo!</p>
+            <p>
+              Deseja realmente excluir o registro{" "}
+              <b className="text-cyan-600">{itemDelete?.firstName}</b>? Para
+              confirmar que você realmente deseja excluir o cliente, digite o
+              nome abaixo
+            </p>
+            <div>
+              <input
+                className="bg-transparent border border-zinc-700 p-3 w-full rounded outline-none focus:border-cyan-500"
+                onChange={onChangeNameConfirmDelete}
+                value={nameConfirmDelete}
+                placeholder="Exemple"
+              />
+            </div>
           </ModalBody>
           <ModalFooter>
             <Button
@@ -83,7 +133,8 @@ export default function Clients() {
               <Button
                 color="danger"
                 className="rounded-lg opacity-80 hover:opacity-100"
-                onClick={() => deleteItem(itemDelete.id)}
+                onClick={() => deleteItem(itemDelete)}
+                disabled={!nameConfirmDelete}
               >
                 Tenho certeza!
               </Button>
